permissions system implemented. todo: add a default user if not exists

+++ src/server/permissions.js
@@ -1,8 +1,14 @@
 const permissionsRouter = require('express').Router()
 const {knex} = require('./db')
+const {decodeJWTPayload} = require('./auth/token')
 // query for practitioner
 
 permissionsRouter.post('/create', async (req, res) => {
+	const {permissions} = decodeJWTPayload(req.headers.token)
+	if (!permissions.includes('edit:link')) {
+		res.sendStatus(401)
+		return
+	}
 	const resp = await knex('practitionerpatients').insert({
 		practitioner_id: req.body.practitionerID,
 		patient_id: req.body.patientID,
@@ -11,6 +17,11 @@ permissionsRouter.post('/create', async (req, res) => {
 })
 
 permissionsRouter.post('/destroy', async (req, res) => {
+	const {permissions} = decodeJWTPayload(req.headers.token)
+	if (!permissions.includes('edit:link')) {
+		res.sendStatus(401)
+		return
+	}
 	const unionDeleted = await knex('practitionerpatients').delete().where({
 		practitioner_id: req.body.practitionerID,
 		patient_id: req.body.patientID,
@@ -18,10 +29,38 @@ permissionsRouter.post('/destroy', async (req, res) => {
 	res.json(unionDeleted)
 })
 
-permissionsRouter.get('/:id', async (req, res) => {
-	const unionTable = await knex('practitionerpatients').select().where({
+permissionsRouter.post('/toggle', async (req, res) => {
+	const decodedToken = decodeJWTPayload(req.headers.token)
+	if (decodedToken.permissions.includes('edit:permissions')) {
+		// if the permission is in the permission set, remove it
+		const index = req.body.set.indexOf(req.body.permission)
+
+		if (index > -1) req.body.set.splice(index)
+		else req.body.set.push(req.body.permission)
+
+		const [row] = await knex('practitioner')
+			.where({practitioner_id: req.body.practitionerID})
+			.update({
+				permissions: JSON.stringify(req.body.set),
+			}).returning('*')
+		delete row.passhash
+		res.send(row)
+	} else {
+		res.sendStatus(401)
+	}
+})
+
+
+permissionsRouter.get('/view/:id', async (req, res) => {
+	const [practitioner] = await knex('practitioner').select().where({
 		practitioner_id: req.params.id,
 	})
+	delete practitioner.passhash
+	res.json(practitioner)
+})
+
+permissionsRouter.get('/relationships/:id', async (req, res) => {
+	const unionTable = await knex('practitionerpatients').select().where({practitioner_id: req.params.id})
 	res.json(unionTable)
 })
